Add CommandContext
to easily extend command handler arguments.  Here it's to prepare
for the support of message tags.  Some command handlers such as
`cmd_join` need to access client-sent tags to relay them to other
clients.  Maybe in the future they will need more data to work on.
To avoid having to change the signature of all handlers every time,
CommandContext has all data related to the execution of the commands.

Nonetheless, command arguments are still out of this struct.  Maybe the
whole message from the client could just be passed in the context and no
other argument would be needed for any command handler.  But I prefer to
precise them explicitly, so that it's clear what are the IRC command
arguments.  Also it avoid having to rebind them in the command handler
to give them explicit names.

Thank you for reading this.

I dare you to read this, future me.  (if you read this go do something
else with your life motherfucker!!!! wake up!)

cringe? no please don't care about it