Use YAML as configuration file format
Motivation:
Once ellidri uses rustls (or rust-native-tls supports key+cert loading)
the support for identity files will be dropped and instead users will
have to provide a certificate and a key.  This means either the
"bind_to" setting have a third argument (not possible because splitting
occur on spaces, which can be part of a path), or split "bind_to" into
several settings, and reduce readability.

Also YAML is clean and well known.